feat(desktop): secure onboarding, recommendation, submit, and dashboard commands
- application service wires encrypted identity vault, SQLite projections, and ranking
- Tauri commands for members, friends, CHECK, SUBMIT, lists, archive, and dashboard
- per-device development encryption key derivation; production OS keyring remains deferred
- event replay preserves existing masked PAN projections
- end-to-end service test proves plaintext PAN exclusion from vault envelope bytes and SQLite
- scanner, clippy, rustfmt, and 115 workspace tests green

+++ crates/ranking/src/lib.rs
@@ -195,7 +195,7 @@ impl RankingAlgorithm for DevRankingAlgorithm {
             ranked.push(RankedIpo {
                 typed_name: name,
                 ranking: index as u32 + 1,
-                score: (10_000 - allocation_bp) as i64,
+                score: 10_000 - allocation_bp,
                 recommended_account_count: count,
                 recommended_allocation_ratio_bp: allocation_bp,
                 skip: false,
